move react component to a separate file
